Fixed #24982 -- Split staticfiles tests into multiple files
